pyproject.toml
[project]
name = "hf-agents-course"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "datasets>=4.0.0",
    "ddgs>=9.5.4",
    "google-genai>=1.32.0",
    "gradio>=5.44.0",
    "httpx>=0.28.1",
    "huggingface-hub[cli,inference]>=0.34.4",
    "langchain-openai>=0.3.31",
    "langgraph>=0.6.6",
    "llama-index-embeddings-huggingface>=0.6.0",
    "llama-index-llms-huggingface-api>=0.6.0",
    "llama-index-llms-lmstudio>=0.4.0",
    "llama-index-retrievers-bm25>=0.6.3",
    "llama-index-tools-duckduckgo>=0.4.0",
    "llama-index-tools-google>=0.6.0",
    "llama-index-vector-stores-chroma>=0.5.2",
    "openai-agents[litellm]>=0.2.10",
    "pillow>=11.3.0",
    "python-dotenv>=1.1.1",
]
